@@ -1,88 +1,32 @@
-/**
- * The first line in the build configuration applies the Android plugin for
- * Gradle to this build and makes the android block available to specify
- * Android-specific build options.
- */
-
-apply plugin: 'com.android.application'
-
-/**
- * The android block is where you configure all your Android-specific
- * build options.
- */
+plugins {
+    id 'com.android.application'
+}
 
 android {
+    namespace = 'com.smorgasbork.hotdeath'
 
-    /**
-     * compileSdkVersion specifies the Android API level Gradle should use to
-     * compile your app. This means your app can use the API features included in
-     * this API level and lower.
-     */
-
-    compileSdkVersion 27
-
-    /**
-     * buildToolsVersion specifies the version of the SDK build tools, command-line
-     * utilities, and compiler that Gradle should use to build your app. You need to
-     * download the build tools using the SDK Manager.
-     *
-     * This property is optional because the plugin uses a recommended version of
-     * the build tools by default.
-     */
-
-    /* buildToolsVersion "29.0.0" */
-
-    /**
-     * The defaultConfig block encapsulates default settings and entries for all
-     * build variants, and can override some attributes in main/AndroidManifest.xml
-     * dynamically from the build system. You can configure product flavors to override
-     * these values for different versions of your app.
-     */
+    compileSdk = 36
 
     defaultConfig {
-
-        /**
-         * applicationId uniquely identifies the package for publishing.
-         * However, your source code should still reference the package name
-         * defined by the package attribute in the main/AndroidManifest.xml file.
-         */
-
-        applicationId 'com.smorgasbork.hotdeath'
-
-        // Defines the minimum API level required to run the app.
-        minSdkVersion 17
-
-        // Specifies the API level used to test the app.
-        targetSdkVersion 28
-
-        // Defines the version number of your app.
-        versionCode 11
-
-        // Defines a user-friendly version name for your app.
-        versionName "1.0.11"
+        applicationId = 'com.smorgasbork.hotdeath'
+        minSdk = 28               // Updated minimum to Android 9 (Pie)
+        targetSdk = 36
+        versionCode = 20
+        versionName = "2.0.00"
     }
 
-    /**
-     * The buildTypes block is where you can configure multiple build types.
-     * By default, the build system defines two build types: debug and release. The
-     * debug build type is not explicitly shown in the default build configuration,
-     * but it includes debugging tools and is signed with the debug key. The release
-     * build type applies Proguard settings and is not signed by default.
-     */
-
     buildTypes {
-
-        /**
-         * By default, Android Studio configures the release build type to enable code
-         * shrinking, using minifyEnabled, and specifies the default Proguard rules file.
-         */
-
         release {
-            minifyEnabled true // Enables code shrinking for the release build type.
-            proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
+            minifyEnabled = true
+            proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
         }
         debug {
             versionNameSuffix = 'g'
         }
+    }
+
+    compileOptions {
+        sourceCompatibility = JavaVersion.VERSION_11
+        targetCompatibility = JavaVersion.VERSION_11
     }
 }
