@@ -2,25 +2,24 @@ package com.smorgasbork.hotdeath;
 
 
 import android.app.Activity;
-import org.json.*;
+import android.app.Dialog;
 import android.content.SharedPreferences;
 import android.content.res.Configuration;
-import android.app.Dialog;
 import android.os.Bundle;
 import android.preference.PreferenceManager;
-import android.view.Menu;
-import android.view.MenuInflater;
-import android.view.MenuItem;
 import android.view.View;
-import android.widget.GridView;
+import android.view.Window;
 import android.widget.AdapterView;
 import android.widget.AdapterView.OnItemClickListener;
-import android.view.*;
-import android.widget.ImageView;
-import android.widget.TextView;
-import android.widget.RelativeLayout;
-import android.widget.RelativeLayout.*;
 import android.widget.Button;
+import android.widget.GridView;
+import android.widget.ImageView;
+import android.widget.RelativeLayout;
+import android.widget.RelativeLayout.LayoutParams;
+import android.widget.TextView;
+
+import org.json.JSONException;
+import org.json.JSONObject;
 
 public class GameActivity extends Activity 
 {
@@ -28,9 +27,6 @@ public class GameActivity extends Activity
 	
 	public static final int STARTUP_MODE_NEW = 1;
 	public static final int STARTUP_MODE_CONTINUE = 2;
-	
-	public static final int DIALOG_CARD_HELP = 0;
-	public static final int DIALOG_CARD_CATALOG = 1;
 	
 	private Dialog m_dlgCardCatalog = null;
 	private Dialog m_dlgCardHelp = null;
@@ -84,7 +80,7 @@ public class GameActivity extends Activity
 	    	JSONObject o;
 	    	try
 	    	{
-		    	o = new JSONObject (s);	    	
+				o = new JSONObject (s);
 				m_game = new Game (o, this, m_go);
 	    	}
 	    	catch (JSONException e)
@@ -105,10 +101,10 @@ public class GameActivity extends Activity
 	    RelativeLayout l = new RelativeLayout (this);
 
 		m_btnFastForward = (Button)getLayoutInflater().inflate(R.layout.action_button, null);
-	    m_btnFastForward.setText(getString(R.string.lbl_fast_forward));
-	    m_btnFastForward.setId(View.generateViewId());
-	    m_btnFastForward.setVisibility(View.INVISIBLE);
-	    m_btnFastForward.setOnClickListener (new View.OnClickListener() {
+		m_btnFastForward.setText(getString(R.string.lbl_fast_forward));
+		m_btnFastForward.setId(View.generateViewId());
+		m_btnFastForward.setVisibility(View.INVISIBLE);
+		m_btnFastForward.setOnClickListener (new View.OnClickListener() {
 			
 			@Override
 			public void onClick(View v) {
@@ -244,12 +240,12 @@ public class GameActivity extends Activity
     protected void onDestroy() {
     	m_game.shutdown ();
     	m_game = null;
-    	m_gt = null;
-       	m_go = null;
-    	
-    	super.onDestroy ();
-    };
-    
+		m_gt = null;
+		m_go = null;
+
+		super.onDestroy ();
+    }
+
     public void showCardHelp ()
     {
     	if (m_dlgCardHelp == null)
